@@ -5,30 +5,29 @@
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build --base-href /webplayer/",
-    "build:prod": "ng build --base-href /webplayer/ --configuration=production",
+    "build:prod": "ng build --prod --base-href /webplayer/",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.2.0",
-    "@angular/cdk": "~7.2.0",
-    "@angular/common": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "~7.2.0",
-    "@angular/forms": "~7.2.0",
-    "@angular/material": "~7.2.0",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
+    "@angular/animations": "~8.2.13",
+    "@angular/cdk": "~8.2.3",
+    "@angular/common": "~8.2.13",
+    "@angular/compiler": "~8.2.13",
+    "@angular/core": "~8.2.13",
+    "@angular/forms": "~8.2.13",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "~8.2.13",
+    "@angular/platform-browser-dynamic": "~8.2.13",
+    "@angular/router": "~8.2.13",
     "@auth0/angular-jwt": "^3.0.1",
     "@ngx-translate/core": "^11.0.1",
     "@ngx-translate/http-loader": "^4.0.0",
     "angular-ng-autocomplete": "^1.1.14",
-    "bootstrap": "^4.4.1",
     "chart.js": "^2.9.3",
-    "core-js": "^2.6.11",
+    "core-js": "^3.6.2",
     "date-fns": "^2.7.0",
     "hammerjs": "^2.0.8",
     "httpclient": "^0.1.0",
@@ -37,30 +36,31 @@
     "materialize-css": "^1.0.0",
     "ngx-materialize": "^6.1.3",
     "ngx-take-until-destroy": "^5.4.0",
-    "rxjs": "6.4.0",
+    "rxjs": "~6.4.0",
     "slick-carousel": "^1.8.1",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1",
+    "ngx-mat-select-search": "^2.1.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "~7.3.3",
-    "@angular/compiler-cli": "~7.2.0",
-    "@angular/language-service": "~7.2.0",
+    "@angular-devkit/build-angular": "~0.803.18",
+    "@angular/cli": "~8.3.18",
+    "@angular/compiler-cli": "~8.2.13",
+    "@angular/language-service": "~8.2.13",
     "@types/node": "~8.9.4",
-    "@types/jasmine": "~2.8.8",
+    "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
+    "codelyzer": "^5.0.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.0.0",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.2.2"
+    "tslint": "~5.15.0",
+    "typescript": "~3.5.3"
   }
 }
