@@ -1,8 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { AutocompleteLibModule } from 'angular-ng-autocomplete';
-
 import { AdminRoutingModule } from './routing/admin-routing.module';
 import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 
@@ -14,7 +12,6 @@ import { CardCreateComponent } from './card/card-create/card-create.component';
 import { TeamCreateComponent } from './team/team-create/team-create.component';
 import { DepartmentCreateComponent } from './department/department-create/department-create.component';
 import { PlayersListComponent } from './player/players-list/players-list.component';
-import { CardListPlayerComponent } from './card/card-list-player/card-list-player.component';
 
 @NgModule({
   declarations: [
@@ -23,19 +20,14 @@ import { CardListPlayerComponent } from './card/card-list-player/card-list-playe
     CardCreateComponent,
     TeamCreateComponent,
     DepartmentCreateComponent,
-    PlayersListComponent,
-    CardListPlayerComponent
+    PlayersListComponent
   ],
   imports: [
     ReactiveFormsModule,
     FormsModule,
     CommonModule,
     SharedModule,
-    AdminRoutingModule,
-    AutocompleteLibModule 
-  ],
-  entryComponents: [
-    CardListPlayerComponent
+    AdminRoutingModule
   ]
 })
 export class AdminModule { }
